+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix issue with removed dnd5e-scriptlets function.
 - Arcane Archer Arrow corrections (2014).
 - Wild Shape and Circle Forms for 2024 Druid. (Thanks to @tminz)
+- New option to save monster tokens to folder names that match actor name. @wasabiprime
 
 # 6.5.6